+++ src/lib/actions/file-system.ts
@@ -1,0 +1,41 @@
+export default class FilesystemActions{
+     public static newTranslation(){
+          console.log("TODO: Implement New Translation Action")
+     }
+     public static openTranslation(){
+          console.log("TODO: Implement Open Translation Action")
+     }
+     public static openRecent(){
+          console.log("TODO: Implement Open Recent translations (by path ofc) Action")
+     }
+     public static saveAll(){
+          console.log("TODO: Implement Save All Action")
+     }
+     public static saveAs(){
+          console.log("TODO: Implement Save as Action")
+     }
+     public static undo(){
+          console.log("TODO: Implement Undo Action")
+     }
+     public static redo(){
+          console.log("TODO: Implement Redo Action")
+     }
+     public static replaceTranslation(){
+          console.log("TODO: Implement Replace translation Action")
+     }
+     public static batchRename(){
+          console.log("TODO: Implement batch rename Action")
+     }
+     public static cut(){
+          console.log("TODO: Implement Cut Action")
+     }
+     public static copy(){
+          console.log("TODO: Implement Copy Action")
+     }
+     public static paste(){
+          console.log("TODO: Implement Paste Action")
+     }
+     public static selectAll(){
+          console.log("TODO: Implement Select All Action")
+     }
+}