+++ src/lib/actions/find.ts
@@ -1,0 +1,14 @@
+export default class FindActions{
+     public static find(query: string){
+          console.log(`TODO: Implement finding with query "${query}"`)
+     }
+     public static findNext(){
+          console.log("TODO: Implement Find Next")
+     }
+     public static findPrev(){
+          console.log("TODO: Implement Find Previous")
+     }
+     public static findMissing(){
+          console.log("TODO: Implement Find Missing keys")
+     }
+}